+++ src/database/operations/authorization/hasChangemakerFieldValuePermission.ts
@@ -1,0 +1,8 @@
+import { generateHasPermissionOperation } from '../generators';
+
+const hasChangemakerFieldValuePermission = generateHasPermissionOperation(
+	'authorization.hasChangemakerFieldValuePermission',
+	'changemakerFieldValueId',
+);
+
+export { hasChangemakerFieldValuePermission };